// x-pack/platform/packages/private/ml/data_frame_analytics_utils/src/types.d.ts
interface Regression {
    /**
     * Dependent variable
     */
    dependent_variable: string;
    /**
     * Training percent
     */
    training_percent: number;
    /**
     * Optional number of top feature importance values
     */
    num_top_feature_importance_values?: number;
    /**
     * Prediction field name
     */
    prediction_field_name?: string;
}
/**
 * Inner interface for classification job configuration options
 */
interface Classification {
    /**
     * Optional class assignment objective
     */
    class_assignment_objective?: string;
    /**
     * Dependent variable
     */
    dependent_variable: string;
    /**
     * Training percent
     */
    training_percent: number;
    /**
     * Optional number of top classes
     */
    num_top_classes?: number;
    /**
     * Optional number of top feature importance values
     */
    num_top_feature_importance_values?: number;
    /**
     * Optional prediction field name
     */
    prediction_field_name?: string;
}
/**
 * Interface for regression analysis job configuation
 */
export interface RegressionAnalysis {
    /**
     * Key spec for interface
     */
    [key: string]: Regression;
    /**
     * Regression options
     */
    regression: Regression;
}
/**
 * Interface for classification job configuation
 */
export interface ClassificationAnalysis {
    /**
     * Key spec for interface
     */
    [key: string]: Classification;
    /**
     * Outlier detection options
     */
    classification: Classification;
}
/**
 * Alias of estypes.MlDataframeAnalysisContainer
 */
export type AnalysisConfig = estypes.MlDataframeAnalysisContainer;